arm: mach-k3: j721e: Move booting of Main R5FSS Core0 to A72 U-Boot
The Main R5FSS Core0 on J721E SoCs is currently booted from R5 SPL
itself to achieve certain product-level early-boot metrics. Move the
booting of this core from R5 SPL to A72 U-Boot in preparation for
the R5 SPL re-architecture code to deal with the System Firmware
functionality split between two different processors.

The env variables are left as is for now, so that, this can potentially
be moved back to be booted from R5 SPL in the future.

Signed-off-by: Suman Anna <s-anna@ti.com>