configs: j7200_evm_a72: Move backup env location to 0x6C0000
The Cypress S28 flash has 256k erase blocks. 0x6A0000 lies in the middle
of an erase block. So to write to 0x6C0000 we would have to erase from
address 0x680000, which is the start of the main env location. Move the
backup location to 0x6C0000 to align it with the start of the next erase
block.

This location used to be for sysfw but with J7200 sysfw is bundled with
R5 SPL.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>