spi: cadence-qspi: Do not use DMA for small reads
A very frequent operation is the Read Status Register command that is
executed repeatedly after doing an erase or page program. On the Cypress
S28HS512T flash this command expects 4 address bytes and so it does not
go via the STIG route. The command reads 2 bytes from the flash. Setting
up DMA for it is more hassle than it is worth. For small reads like
this, the speed gain, if any, will not be worth the extra overhead.

So do not use DMA for reads smaller than 16 bytes. This should cover
polling operations like SR polling.

It also fixes an issue on J721E and J7200 with the Cypress S28HS512T
flash where DMA would eventually freeze when writing or erasing because
of the constant bombardment of small, frequent reads.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>