TEMP: drivers: ram: k3-am64: Introduce ddrss driver
The AM64x DDR subsystem comprises DDR controller, DDR PHY and wrapper
logic to integrate these blocks in the device. The DDR subsystem is
used to provide an interface to external SDRAM devices which can be
utilized for storing program or data. Introduce support for the
DDR controller and DDR phy within the DDR subsystem.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>