arm: dts: k3-j7200: Set initial TX start to 18
TX = 16 is on the border of the range of valid TX values for this board
with the Cypress S28 flash. As a result, the PHY tuning sometimes fails
to find rxlow at that point, causing the algorithm to iterate over the
TX range taking a hit on the tuning time.

TX = 18 is much more stable in comparison. Tuning consistently succeeds
with it. Use it as the initial value to stop the PHY tuning from taking
longer.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>