ram: k3-am64: Configure DDR PLL before controller start for DDR4
DDR4 does not require the frequency handshaking needed for LPDDR4, and
attempting to just statically configure the PLL at this time for DDR4
seems to produce marginal stability. To remove any risk of DDR4 getting
stuck initializing, configure the clock before attempting to start the
controller for DDR4 and leave any frequency handshaking to LPDDR4 only.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Tested-by: Vignesh Raghavendra <vigneshr@ti.com>