arm: dts: k3-am65: Add remoteproc-name to ICSSG remote processors
If a remoteproc-name DT property is not present then the u-boot
remoteproc class driver uses the node name as remote processor name.

As the remote processor node names are not unique between different
ICSSG instances it causes probe failure for subsequent ICSSG instances.

To resolve this we specify a unique name using the remoteproc-name
DT property.

Signed-off-by: Roger Quadros <rogerq@ti.com>