arm: dts: k3-am65: Add tag to disable UHS modes in SR 1.0
The 4 bit interface in am65 sr1.0 cannot be supported at 3.3V or 1.8V
because of erratas i2025 and i2026. However, SD card is the primary
boot mode for development usecases.

Continue to enable SD card in SR1.0 and disable UHS modes in it to
minimise any ageing issues happening because of the erratas.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Reviewed-by: Kishon Vijay Abraham I <kishon@ti.com>