clk: k3: Minimize rounding error when programming dividers
The clk-divider driver rounds up extremely aggressively, to the point
that any remainder from divider calculation at all results in a div + 1
value being used. This can result is a major delta in frequency from
what is expected.

We can account for this by testing the diff between expected and
requested frequencies and if wrose than 5% difference, request rate + 1
and see if the delta is better. If so, use the new rate, and if not,
return to the original value. This can result in multiple rate changes
when attempting to program a clock, however it should occur rarely, when
the requested rate is far off from the rate received.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>