clk: k3: Extend rounding error correction when programming dividers
The clk-divider driver rounds up extremely aggressively, to the point
that any remainder from divider calculation at all results in a div + 1
value being used. This can result is a major delta in frequency from
what is expected.

This was initially addressed in patch
commit b96d9b5a3f8b ("clk: k3: Minimize rounding error when programming dividers")
to account for off by one errors resulting in extreme rounding,
but a wider range must be considered.

Continue to account for this by testing the diff between expected and
requested frequencies and if worse than 5% difference, do the division
manually and take the remainder from the calculation, add it to the
requested rate, and see if the delta is better. If so, use the new rate,
and if not, return to the original value.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>