clk: k3: Change parent PLL frequency determination
When set rate is called on a divider clock, the clk-k3 driver detects if
programming the divider alone gives worse than 50% accuracy and if so,
attempts to program a new frequency using both the divider itself and
the parent clock. For an HSDIV the parent is a PLL, so some logic is
needed to determine what the PLL should output.

Currently the driver takes the MAX PLL frequency and divides it by the
rate to both a divider value to use along with the PLL target freq to
program. This gives poor accuracy and prioritizes programming PLL MAX
for the PLL target freq.

Instead, first check if the divider value must be one and program the
freq directly into the PLL, otherwise start multiplying the requested
frequency by 2 up to the maximum value for the divider to find a PLL
target frequency that falls within acceptable VCO range.

This gives better accuracy for the requested frequency and also more
reasonable VCO values for the PLL.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>