arm: mach-k3: sysfw-loader: Pass RM boardcfg to TIFS after storing for DM
To allow use of TIFS RM services from SPL, a change has been made to
pass the RM boardcfg directly to TIFS from SPL before storing the RM
boardcfg in memory for DM to consume once it boots. However, on HS
silicon, the TIFS RM boardcfg processing modifies the passed blob in
place, which is passed as signing cert + RM blon, and is left as just
the the raw RM blob. This then gets copied to the DM boardcfg shared
memory, and DM fails.

DM expects to have the signing cert + blob shared with it, so to allow
this to happen, copy the signing cert + RM blob first, and move sending
to TIFS after. The boardcfg gets modified in its original location but
we have already copied the cert and blob and they are passed to DM
unmodified.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>